shift response handling responsibility to client
