+++ lib/iban_calculator/response.rb
@@ -1,0 +1,6 @@
+module IbanCalculator
+  class Response
+    def initialize(response)
+    end
+  end
+end
